refactor: moved parameter structs out into folder named by the spec they come from
