+++ src/rdm/parameter/e120.rs
@@ -1,395 +1,10 @@
-use super::{RdmError, SubDeviceId};
-use core::{
-    fmt,
-    net::{Ipv4Addr, Ipv6Addr},
-    result::Result,
-};
+use super::{decode_string_bytes, RdmError, SubDeviceId};
+use core::fmt;
 
 #[cfg(not(feature = "alloc"))]
 use core::str::FromStr;
 #[cfg(not(feature = "alloc"))]
-use heapless::{String, Vec};
-
-#[cfg(feature = "alloc")]
-pub fn decode_string_bytes(bytes: &[u8]) -> Result<String, RdmError> {
-    let utf8 = String::from_utf8_lossy(bytes);
-
-    if utf8.contains(char::from(0)) {
-        Ok(utf8.split_once(char::from(0)).unwrap().0.to_string())
-    } else {
-        Ok(utf8.to_string())
-    }
-}
-
-#[cfg(not(feature = "alloc"))]
-pub fn decode_string_bytes<const N: usize>(bytes: &[u8]) -> Result<String<N>, RdmError> {
-    let utf8 = String::<N>::from_utf8(Vec::<u8, N>::from_slice(bytes).unwrap())?;
-
-    if utf8.contains(char::from(0)) {
-        Ok(String::<N>::from_str(utf8.split_once(char::from(0)).unwrap().0).unwrap())
-    } else {
-        Ok(utf8)
-    }
-}
-
-#[non_exhaustive]
-#[derive(Copy, Clone, Debug, PartialEq)]
-pub enum ParameterId {
-    // E1.20
-    DiscUniqueBranch,
-    DiscMute,
-    DiscUnMute,
-    ProxiedDevices,
-    ProxiedDeviceCount,
-    CommsStatus,
-    QueuedMessage,
-    StatusMessages,
-    StatusIdDescription,
-    ClearStatusId,
-    SubDeviceIdStatusReportThreshold,
-    SupportedParameters,
-    ParameterDescription,
-    DeviceInfo,
-    ProductDetailIdList,
-    DeviceModelDescription,
-    ManufacturerLabel,
-    DeviceLabel,
-    FactoryDefaults,
-    LanguageCapabilities,
-    Language,
-    SoftwareVersionLabel,
-    BootSoftwareVersionId,
-    BootSoftwareVersionLabel,
-    DmxPersonality,
-    DmxPersonalityDescription,
-    DmxStartAddress,
-    SlotInfo,
-    SlotDescription,
-    DefaultSlotValue,
-    SensorDefinition,
-    SensorValue,
-    RecordSensors,
-    DeviceHours,
-    LampHours,
-    LampStrikes,
-    LampState,
-    LampOnMode,
-    DevicePowerCycles,
-    DisplayInvert,
-    DisplayLevel,
-    PanInvert,
-    TiltInvert,
-    PanTiltSwap,
-    RealTimeClock,
-    IdentifyDevice,
-    ResetDevice,
-    PowerState,
-    PerformSelfTest,
-    SelfTestDescription,
-    CapturePreset,
-    PresetPlayback,
-    // E1.37-1
-    DmxBlockAddress,
-    DmxFailMode,
-    DmxStartupMode,
-    DimmerInfo,
-    MinimumLevel,
-    MaximumLevel,
-    Curve,
-    CurveDescription,
-    OutputResponseTime,
-    OutputResponseTimeDescription,
-    ModulationFrequency,
-    ModulationFrequencyDescription,
-    BurnIn,
-    LockPin,
-    LockState,
-    LockStateDescription,
-    IdentifyMode,
-    PresetInfo,
-    PresetStatus,
-    PresetMergeMode,
-    PowerOnSelfTest,
-    // E1.37-2
-    ListInterfaces,
-    InterfaceLabel,
-    InterfaceHardwareAddressType1,
-    IpV4DhcpMode,
-    IpV4ZeroConfMode,
-    IpV4CurrentAddress,
-    IpV4StaticAddress,
-    InterfaceRenewDhcp,
-    InterfaceReleaseDhcp,
-    InterfaceApplyConfiguration,
-    IpV4DefaultRoute,
-    DnsIpV4NameServer,
-    DnsHostName,
-    DnsDomainName,
-    // E1.37-7
-    EndpointList,
-    EndpointListChange,
-    IdentifyEndpoint,
-    EndpointToUniverse,
-    EndpointMode,
-    EndpointLabel,
-    RdmTrafficEnable,
-    DiscoveryState,
-    BackgroundDiscovery,
-    EndpointTiming,
-    EndpointTimingDescription,
-    EndpointResponders,
-    EndpointResponderListChange,
-    BindingControlFields,
-    BackgroundQueuedStatusPolicy,
-    BackgroundQueuedStatusPolicyDescription,
-    // E1.33
-    ComponentScope,
-    SearchDomain,
-    TcpCommsStatus,
-    BrokerStatus,
-    // use for unsupported standard and manufacturer specific parameters
-    RawParameterId(u16),
-}
-
-impl From<u16> for ParameterId {
-    fn from(value: u16) -> Self {
-        match value {
-            // E1.20
-            0x0001 => Self::DiscUniqueBranch,
-            0x0002 => Self::DiscMute,
-            0x0003 => Self::DiscUnMute,
-            0x0010 => Self::ProxiedDevices,
-            0x0011 => Self::ProxiedDeviceCount,
-            0x0015 => Self::CommsStatus,
-            0x0020 => Self::QueuedMessage,
-            0x0030 => Self::StatusMessages,
-            0x0031 => Self::StatusIdDescription,
-            0x0032 => Self::ClearStatusId,
-            0x0033 => Self::SubDeviceIdStatusReportThreshold,
-            0x0050 => Self::SupportedParameters,
-            0x0051 => Self::ParameterDescription,
-            0x0060 => Self::DeviceInfo,
-            0x0070 => Self::ProductDetailIdList,
-            0x0080 => Self::DeviceModelDescription,
-            0x0081 => Self::ManufacturerLabel,
-            0x0082 => Self::DeviceLabel,
-            0x0090 => Self::FactoryDefaults,
-            0x00a0 => Self::LanguageCapabilities,
-            0x00b0 => Self::Language,
-            0x00c0 => Self::SoftwareVersionLabel,
-            0x00c1 => Self::BootSoftwareVersionId,
-            0x00c2 => Self::BootSoftwareVersionLabel,
-            0x00e0 => Self::DmxPersonality,
-            0x00e1 => Self::DmxPersonalityDescription,
-            0x00f0 => Self::DmxStartAddress,
-            0x0120 => Self::SlotInfo,
-            0x0121 => Self::SlotDescription,
-            0x0122 => Self::DefaultSlotValue,
-            0x0200 => Self::SensorDefinition,
-            0x0201 => Self::SensorValue,
-            0x0202 => Self::RecordSensors,
-            0x0400 => Self::DeviceHours,
-            0x0401 => Self::LampHours,
-            0x0402 => Self::LampStrikes,
-            0x0403 => Self::LampState,
-            0x0404 => Self::LampOnMode,
-            0x0405 => Self::DevicePowerCycles,
-            0x0500 => Self::DisplayInvert,
-            0x0501 => Self::DisplayLevel,
-            0x0600 => Self::PanInvert,
-            0x0601 => Self::TiltInvert,
-            0x0602 => Self::PanTiltSwap,
-            0x0603 => Self::RealTimeClock,
-            0x1000 => Self::IdentifyDevice,
-            0x1001 => Self::ResetDevice,
-            0x1010 => Self::PowerState,
-            0x1020 => Self::PerformSelfTest,
-            0x1021 => Self::SelfTestDescription,
-            0x1030 => Self::CapturePreset,
-            0x1031 => Self::PresetPlayback,
-            // E1.37-1
-            0x0140 => Self::DmxBlockAddress,
-            0x0141 => Self::DmxFailMode,
-            0x0142 => Self::DmxStartupMode,
-            0x0340 => Self::DimmerInfo,
-            0x0341 => Self::MinimumLevel,
-            0x0342 => Self::MaximumLevel,
-            0x0343 => Self::Curve,
-            0x0344 => Self::CurveDescription,
-            0x0345 => Self::OutputResponseTime,
-            0x0346 => Self::OutputResponseTimeDescription,
-            0x0347 => Self::ModulationFrequency,
-            0x0348 => Self::ModulationFrequencyDescription,
-            0x0440 => Self::BurnIn,
-            0x0640 => Self::LockPin,
-            0x0641 => Self::LockState,
-            0x0642 => Self::LockStateDescription,
-            0x1040 => Self::IdentifyMode,
-            0x1041 => Self::PresetInfo,
-            0x1042 => Self::PresetStatus,
-            0x1043 => Self::PresetMergeMode,
-            0x1044 => Self::PowerOnSelfTest,
-            // E1.37-2
-            0x0700 => Self::ListInterfaces,
-            0x0701 => Self::InterfaceLabel,
-            0x0702 => Self::InterfaceHardwareAddressType1,
-            0x0703 => Self::IpV4DhcpMode,
-            0x0704 => Self::IpV4ZeroConfMode,
-            0x0705 => Self::IpV4CurrentAddress,
-            0x0706 => Self::IpV4StaticAddress,
-            0x0707 => Self::InterfaceRenewDhcp,
-            0x0708 => Self::InterfaceReleaseDhcp,
-            0x0709 => Self::InterfaceApplyConfiguration,
-            0x070a => Self::IpV4DefaultRoute,
-            0x070b => Self::DnsIpV4NameServer,
-            0x070c => Self::DnsHostName,
-            0x070d => Self::DnsDomainName,
-            // E1.37-7
-            0x0900 => Self::EndpointList,
-            0x0901 => Self::EndpointListChange,
-            0x0902 => Self::IdentifyEndpoint,
-            0x0903 => Self::EndpointToUniverse,
-            0x0904 => Self::EndpointMode,
-            0x0905 => Self::EndpointLabel,
-            0x0906 => Self::RdmTrafficEnable,
-            0x0907 => Self::DiscoveryState,
-            0x0908 => Self::BackgroundDiscovery,
-            0x0909 => Self::EndpointTiming,
-            0x090a => Self::EndpointTimingDescription,
-            0x090b => Self::EndpointResponders,
-            0x090c => Self::EndpointResponderListChange,
-            0x090d => Self::BindingControlFields,
-            0x090e => Self::BackgroundQueuedStatusPolicy,
-            0x090f => Self::BackgroundQueuedStatusPolicyDescription,
-            // E1.33
-            0x8000 => Self::ComponentScope,
-            0x8001 => Self::SearchDomain,
-            0x8002 => Self::TcpCommsStatus,
-            0x8003 => Self::BrokerStatus,
-            n => Self::RawParameterId(n),
-        }
-    }
-}
-
-impl From<ParameterId> for u16 {
-    fn from(value: ParameterId) -> Self {
-        match value {
-            // E1.20
-            ParameterId::DiscUniqueBranch => 0x0001,
-            ParameterId::DiscMute => 0x0002,
-            ParameterId::DiscUnMute => 0x0003,
-            ParameterId::ProxiedDevices => 0x0010,
-            ParameterId::ProxiedDeviceCount => 0x0011,
-            ParameterId::CommsStatus => 0x0015,
-            ParameterId::QueuedMessage => 0x0020,
-            ParameterId::StatusMessages => 0x0030,
-            ParameterId::StatusIdDescription => 0x0031,
-            ParameterId::ClearStatusId => 0x0032,
-            ParameterId::SubDeviceIdStatusReportThreshold => 0x0033,
-            ParameterId::SupportedParameters => 0x0050,
-            ParameterId::ParameterDescription => 0x0051,
-            ParameterId::DeviceInfo => 0x0060,
-            ParameterId::ProductDetailIdList => 0x0070,
-            ParameterId::DeviceModelDescription => 0x0080,
-            ParameterId::ManufacturerLabel => 0x0081,
-            ParameterId::DeviceLabel => 0x0082,
-            ParameterId::FactoryDefaults => 0x0090,
-            ParameterId::LanguageCapabilities => 0x00a0,
-            ParameterId::Language => 0x00b0,
-            ParameterId::SoftwareVersionLabel => 0x00c0,
-            ParameterId::BootSoftwareVersionId => 0x00c1,
-            ParameterId::BootSoftwareVersionLabel => 0x00c2,
-            ParameterId::DmxPersonality => 0x00e0,
-            ParameterId::DmxPersonalityDescription => 0x00e1,
-            ParameterId::DmxStartAddress => 0x00f0,
-            ParameterId::SlotInfo => 0x0120,
-            ParameterId::SlotDescription => 0x0121,
-            ParameterId::DefaultSlotValue => 0x0122,
-            ParameterId::SensorDefinition => 0x0200,
-            ParameterId::SensorValue => 0x0201,
-            ParameterId::RecordSensors => 0x0202,
-            ParameterId::DeviceHours => 0x0400,
-            ParameterId::LampHours => 0x0401,
-            ParameterId::LampStrikes => 0x0402,
-            ParameterId::LampState => 0x0403,
-            ParameterId::LampOnMode => 0x0404,
-            ParameterId::DevicePowerCycles => 0x0405,
-            ParameterId::DisplayInvert => 0x0500,
-            ParameterId::DisplayLevel => 0x0501,
-            ParameterId::PanInvert => 0x0600,
-            ParameterId::TiltInvert => 0x0601,
-            ParameterId::PanTiltSwap => 0x0602,
-            ParameterId::RealTimeClock => 0x0603,
-            ParameterId::IdentifyDevice => 0x1000,
-            ParameterId::ResetDevice => 0x1001,
-            ParameterId::PowerState => 0x1010,
-            ParameterId::PerformSelfTest => 0x1020,
-            ParameterId::SelfTestDescription => 0x1021,
-            ParameterId::CapturePreset => 0x1030,
-            ParameterId::PresetPlayback => 0x1031,
-            // E1.37-1
-            ParameterId::DmxBlockAddress => 0x0140,
-            ParameterId::DmxFailMode => 0x0141,
-            ParameterId::DmxStartupMode => 0x0142,
-            ParameterId::DimmerInfo => 0x0340,
-            ParameterId::MinimumLevel => 0x0341,
-            ParameterId::MaximumLevel => 0x0342,
-            ParameterId::Curve => 0x0343,
-            ParameterId::CurveDescription => 0x0344,
-            ParameterId::OutputResponseTime => 0x0345,
-            ParameterId::OutputResponseTimeDescription => 0x0346,
-            ParameterId::ModulationFrequency => 0x0347,
-            ParameterId::ModulationFrequencyDescription => 0x0348,
-            ParameterId::BurnIn => 0x0440,
-            ParameterId::LockPin => 0x0640,
-            ParameterId::LockState => 0x0641,
-            ParameterId::LockStateDescription => 0x0642,
-            ParameterId::IdentifyMode => 0x1040,
-            ParameterId::PresetInfo => 0x1041,
-            ParameterId::PresetStatus => 0x1042,
-            ParameterId::PresetMergeMode => 0x1043,
-            ParameterId::PowerOnSelfTest => 0x1044,
-            // E1.37-2
-            ParameterId::ListInterfaces => 0x0700,
-            ParameterId::InterfaceLabel => 0x0701,
-            ParameterId::InterfaceHardwareAddressType1 => 0x0702,
-            ParameterId::IpV4DhcpMode => 0x0703,
-            ParameterId::IpV4ZeroConfMode => 0x0704,
-            ParameterId::IpV4CurrentAddress => 0x0705,
-            ParameterId::IpV4StaticAddress => 0x0706,
-            ParameterId::InterfaceRenewDhcp => 0x0707,
-            ParameterId::InterfaceReleaseDhcp => 0x0708,
-            ParameterId::InterfaceApplyConfiguration => 0x0709,
-            ParameterId::IpV4DefaultRoute => 0x070a,
-            ParameterId::DnsIpV4NameServer => 0x070b,
-            ParameterId::DnsHostName => 0x070c,
-            ParameterId::DnsDomainName => 0x070d,
-            // E1.37-7
-            ParameterId::EndpointList => 0x0900,
-            ParameterId::EndpointListChange => 0x0901,
-            ParameterId::IdentifyEndpoint => 0x0902,
-            ParameterId::EndpointToUniverse => 0x0903,
-            ParameterId::EndpointMode => 0x0904,
-            ParameterId::EndpointLabel => 0x0905,
-            ParameterId::RdmTrafficEnable => 0x0906,
-            ParameterId::DiscoveryState => 0x0907,
-            ParameterId::BackgroundDiscovery => 0x0908,
-            ParameterId::EndpointTiming => 0x0909,
-            ParameterId::EndpointTimingDescription => 0x090a,
-            ParameterId::EndpointResponders => 0x090b,
-            ParameterId::EndpointResponderListChange => 0x090c,
-            ParameterId::BindingControlFields => 0x090d,
-            ParameterId::BackgroundQueuedStatusPolicy => 0x090e,
-            ParameterId::BackgroundQueuedStatusPolicyDescription => 0x090f,
-            // E1.33
-            ParameterId::ComponentScope => 0x0800,
-            ParameterId::SearchDomain => 0x0801,
-            ParameterId::TcpCommsStatus => 0x0802,
-            ParameterId::BrokerStatus => 0x0803,
-            ParameterId::RawParameterId(pid) => pid,
-        }
-    }
-}
+use heapless::String;
 
 #[derive(Copy, Clone, Debug, PartialEq, Eq)]
 pub struct ProtocolVersion {
@@ -2161,664 +1776,5 @@ impl SensorValue {
             highest_detected_value,
             recorded_value,
         }
-    }
-}
-
-#[derive(Copy, Clone, Debug, PartialEq)]
-pub enum IdentifyMode {
-    Quiet = 0x00,
-    Loud = 0xff,
-}
-
-impl TryFrom<u8> for IdentifyMode {
-    type Error = RdmError;
-
-    fn try_from(value: u8) -> Result<Self, Self::Error> {
-        match value {
-            0x00 => Ok(Self::Quiet),
-            0xff => Ok(Self::Loud),
-            value => Err(RdmError::InvalidIdentifyMode(value)),
-        }
-    }
-}
-
-#[derive(Copy, Clone, Debug, PartialEq)]
-pub enum PresetProgrammed {
-    NotProgrammed = 0x00,
-    Programmed = 0x01,
-    ReadOnly = 0x02,
-}
-
-impl TryFrom<u8> for PresetProgrammed {
-    type Error = RdmError;
-
-    fn try_from(value: u8) -> Result<Self, Self::Error> {
-        match value {
-            0x00 => Ok(Self::NotProgrammed),
-            0x01 => Ok(Self::Programmed),
-            0x02 => Ok(Self::ReadOnly),
-            value => Err(RdmError::InvalidPresetProgrammed(value)),
-        }
-    }
-}
-
-#[derive(Copy, Clone, Debug, PartialEq)]
-pub enum MergeMode {
-    Default = 0x00,
-    Htp = 0x01,
-    Ltp = 0x02,
-    DmxOnly = 0x03,
-    Other = 0xff,
-}
-
-impl TryFrom<u8> for MergeMode {
-    type Error = RdmError;
-
-    fn try_from(value: u8) -> Result<Self, Self::Error> {
-        match value {
-            0x00 => Ok(Self::Default),
-            0x01 => Ok(Self::Htp),
-            0x02 => Ok(Self::Ltp),
-            0x03 => Ok(Self::DmxOnly),
-            0xff => Ok(Self::Other),
-            value => Err(RdmError::InvalidMergeMode(value)),
-        }
-    }
-}
-
-#[derive(Copy, Clone, Debug, PartialEq)]
-pub struct PinCode(pub u16);
-
-impl TryFrom<u16> for PinCode {
-    type Error = RdmError;
-
-    fn try_from(value: u16) -> Result<Self, Self::Error> {
-        if value > 9999 {
-            Err(RdmError::InvalidPinCode(value))
-        } else {
-            Ok(Self(value))
-        }
-    }
-}
-
-#[derive(Copy, Clone, Debug, PartialEq)]
-pub enum SupportedTimes {
-    NotSupported,
-    Time(u16),
-}
-
-impl From<u16> for SupportedTimes {
-    fn from(value: u16) -> Self {
-        match value {
-            0xffff => Self::NotSupported,
-            value => Self::Time(value),
-        }
-    }
-}
-
-impl From<SupportedTimes> for u16 {
-    fn from(value: SupportedTimes) -> u16 {
-        match value {
-            SupportedTimes::NotSupported => 0xffff,
-            SupportedTimes::Time(value) => value,
-        }
-    }
-}
-
-#[derive(Copy, Clone, Debug, PartialEq)]
-pub enum TimeMode {
-    Infinite,
-    TenthOfSeconds(u16),
-}
-
-impl From<u16> for TimeMode {
-    fn from(value: u16) -> Self {
-        match value {
-            0xffff => Self::Infinite,
-            value => Self::TenthOfSeconds(value),
-        }
-    }
-}
-
-impl From<TimeMode> for u16 {
-    fn from(value: TimeMode) -> u16 {
-        match value {
-            TimeMode::Infinite => 0xffff,
-            TimeMode::TenthOfSeconds(value) => value,
-        }
-    }
-}
-
-#[derive(Copy, Clone, Debug, PartialEq)]
-pub enum DhcpMode {
-    Inactive = 0x00,
-    Active = 0x01,
-    Unknown = 0x02,
-}
-
-impl TryFrom<u8> for DhcpMode {
-    type Error = RdmError;
-
-    fn try_from(value: u8) -> Result<Self, Self::Error> {
-        match value {
-            0x00 => Ok(Self::Inactive),
-            0x01 => Ok(Self::Active),
-            0x02 => Ok(Self::Unknown),
-            value => Err(RdmError::InvalidDhcpMode(value)),
-        }
-    }
-}
-
-#[derive(Copy, Clone, Debug, PartialEq)]
-pub enum Ipv4Address {
-    Unconfigured,
-    Configured(Ipv4Addr),
-}
-
-impl From<Ipv4Addr> for Ipv4Address {
-    fn from(value: Ipv4Addr) -> Self {
-        Self::Configured(value)
-    }
-}
-
-impl From<u32> for Ipv4Address {
-    fn from(value: u32) -> Self {
-        if value == 0 {
-            Self::Unconfigured
-        } else {
-            Self::Configured(Ipv4Addr::from(value))
-        }
-    }
-}
-
-impl From<[u8; 4]> for Ipv4Address {
-    fn from(value: [u8; 4]) -> Self {
-        if value == [0, 0, 0, 0] {
-            Self::Unconfigured
-        } else {
-            Self::Configured(Ipv4Addr::from(value))
-        }
-    }
-}
-
-impl From<Ipv4Address> for [u8; 4] {
-    fn from(value: Ipv4Address) -> [u8; 4] {
-        match value {
-            Ipv4Address::Unconfigured => [0, 0, 0, 0],
-            Ipv4Address::Configured(ip) => ip.octets(),
-        }
-    }
-}
-
-impl From<Ipv4Address> for u32 {
-    fn from(value: Ipv4Address) -> u32 {
-        match value {
-            Ipv4Address::Unconfigured => 0,
-            Ipv4Address::Configured(ip) => ip.to_bits(),
-        }
-    }
-}
-
-#[derive(Copy, Clone, Debug, PartialEq)]
-pub enum Ipv6Address {
-    Unconfigured,
-    Configured(Ipv6Addr),
-}
-
-impl From<Ipv6Addr> for Ipv6Address {
-    fn from(value: Ipv6Addr) -> Self {
-        Self::Configured(value)
-    }
-}
-
-impl From<u128> for Ipv6Address {
-    fn from(value: u128) -> Self {
-        if value == 0 {
-            Self::Unconfigured
-        } else {
-            Self::Configured(Ipv6Addr::from(value))
-        }
-    }
-}
-
-impl From<[u8; 16]> for Ipv6Address {
-    fn from(value: [u8; 16]) -> Self {
-        if value == [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0] {
-            Self::Unconfigured
-        } else {
-            Self::Configured(Ipv6Addr::from(value))
-        }
-    }
-}
-
-impl From<Ipv6Address> for [u8; 16] {
-    fn from(value: Ipv6Address) -> [u8; 16] {
-        match value {
-            Ipv6Address::Unconfigured => [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-            Ipv6Address::Configured(ip) => ip.octets(),
-        }
-    }
-}
-
-impl From<Ipv6Address> for u128 {
-    fn from(value: Ipv6Address) -> u128 {
-        match value {
-            Ipv6Address::Unconfigured => 0,
-            Ipv6Address::Configured(ip) => ip.to_bits(),
-        }
-    }
-}
-
-#[derive(Copy, Clone, Debug, PartialEq)]
-pub enum Ipv4Route {
-    NoDefault,
-    Configured(Ipv4Addr),
-}
-
-impl From<Ipv4Addr> for Ipv4Route {
-    fn from(value: Ipv4Addr) -> Self {
-        Self::Configured(value)
-    }
-}
-
-impl From<u32> for Ipv4Route {
-    fn from(value: u32) -> Self {
-        if value == 0 {
-            Self::NoDefault
-        } else {
-            Self::Configured(Ipv4Addr::from(value))
-        }
-    }
-}
-
-impl From<[u8; 4]> for Ipv4Route {
-    fn from(value: [u8; 4]) -> Self {
-        if value == [0, 0, 0, 0] {
-            Self::NoDefault
-        } else {
-            Self::Configured(Ipv4Addr::from(value))
-        }
-    }
-}
-
-impl From<Ipv4Route> for [u8; 4] {
-    fn from(value: Ipv4Route) -> [u8; 4] {
-        match value {
-            Ipv4Route::NoDefault => [0, 0, 0, 0],
-            Ipv4Route::Configured(ip) => ip.octets(),
-        }
-    }
-}
-
-impl From<Ipv4Route> for u32 {
-    fn from(value: Ipv4Route) -> u32 {
-        match value {
-            Ipv4Route::NoDefault => 0,
-            Ipv4Route::Configured(ip) => ip.to_bits(),
-        }
-    }
-}
-
-// Hardware types are defined by the IANA:
-// https://www.iana.org/assignments/arp-parameters/arp-parameters.xhtml
-#[derive(Copy, Clone, Debug, PartialEq)]
-pub enum HardwareType {
-    Reserved(u16),
-    Ethernet,
-    ExperimentEthernet,
-    AmateurRadioAx25,
-    ProteonPronetTokenRing,
-    Chaos,
-    Ieee802Networks,
-    Arcnet,
-    Hyperchannel,
-    Lanstar,
-    AutonetShortAddress,
-    LocalTalk,
-    LocalNet,
-    UltraLink,
-    SMDS,
-    FrameRelay,
-    ATM,
-    HDLC,
-    FibreChannel,
-    ATMLogical,
-    SerialLine,
-    ATMPhysical,
-    MilStd188220,
-    Metricom,
-    IEEE1394,
-    MAPOS,
-    Twinaxial,
-    EUI64,
-    HIPARP,
-    IPAndARPOverISO,
-    ARPSec,
-    IPsecTunnel,
-    InfiniBand,
-    TIA102,
-    Wiegand,
-    PureIP,
-    HwExp1,
-    Hf1,
-    UnifiedBus,
-    HwExp2,
-    AEthernet,
-    Unknown(u16),
-}
-
-impl From<u16> for HardwareType {
-    fn from(value: u16) -> Self {
-        match value {
-            val @ (0 | 65535) => Self::Reserved(val),
-            1 => Self::Ethernet,
-            2 => Self::ExperimentEthernet,
-            3 => Self::AmateurRadioAx25,
-            4 => Self::ProteonPronetTokenRing,
-            5 => Self::Chaos,
-            6 => Self::Ieee802Networks,
-            7 => Self::Arcnet,
-            8 => Self::Hyperchannel,
-            9 => Self::Lanstar,
-            10 => Self::AutonetShortAddress,
-            11 => Self::LocalTalk,
-            12 => Self::LocalNet,
-            13 => Self::UltraLink,
-            14 => Self::SMDS,
-            15 => Self::FrameRelay,
-            16 => Self::ATM,
-            17 => Self::HDLC,
-            18 => Self::FibreChannel,
-            19 => Self::ATMLogical,
-            20 => Self::SerialLine,
-            21 => Self::ATMPhysical,
-            22 => Self::MilStd188220,
-            23 => Self::Metricom,
-            24 => Self::IEEE1394,
-            25 => Self::MAPOS,
-            26 => Self::Twinaxial,
-            27 => Self::EUI64,
-            28 => Self::HIPARP,
-            29 => Self::IPAndARPOverISO,
-            30 => Self::ARPSec,
-            31 => Self::IPsecTunnel,
-            32 => Self::InfiniBand,
-            33 => Self::TIA102,
-            34 => Self::Wiegand,
-            35 => Self::PureIP,
-            36 => Self::HwExp1,
-            37 => Self::Hf1,
-            38 => Self::UnifiedBus,
-            256 => Self::HwExp2,
-            257 => Self::AEthernet,
-            value => Self::Unknown(value),
-        }
-    }
-}
-
-impl From<HardwareType> for u16 {
-    fn from(value: HardwareType) -> Self {
-        match value {
-            HardwareType::Reserved(val) => val,
-            HardwareType::Ethernet => 1,
-            HardwareType::ExperimentEthernet => 2,
-            HardwareType::AmateurRadioAx25 => 3,
-            HardwareType::ProteonPronetTokenRing => 4,
-            HardwareType::Chaos => 5,
-            HardwareType::Ieee802Networks => 6,
-            HardwareType::Arcnet => 7,
-            HardwareType::Hyperchannel => 8,
-            HardwareType::Lanstar => 9,
-            HardwareType::AutonetShortAddress => 10,
-            HardwareType::LocalTalk => 11,
-            HardwareType::LocalNet => 12,
-            HardwareType::UltraLink => 13,
-            HardwareType::SMDS => 14,
-            HardwareType::FrameRelay => 15,
-            HardwareType::ATM => 16,
-            HardwareType::HDLC => 17,
-            HardwareType::FibreChannel => 18,
-            HardwareType::ATMLogical => 19,
-            HardwareType::SerialLine => 20,
-            HardwareType::ATMPhysical => 21,
-            HardwareType::MilStd188220 => 22,
-            HardwareType::Metricom => 23,
-            HardwareType::IEEE1394 => 24,
-            HardwareType::MAPOS => 25,
-            HardwareType::Twinaxial => 26,
-            HardwareType::EUI64 => 27,
-            HardwareType::HIPARP => 28,
-            HardwareType::IPAndARPOverISO => 29,
-            HardwareType::ARPSec => 30,
-            HardwareType::IPsecTunnel => 31,
-            HardwareType::InfiniBand => 32,
-            HardwareType::TIA102 => 33,
-            HardwareType::Wiegand => 34,
-            HardwareType::PureIP => 35,
-            HardwareType::HwExp1 => 36,
-            HardwareType::Hf1 => 37,
-            HardwareType::UnifiedBus => 38,
-            HardwareType::HwExp2 => 256,
-            HardwareType::AEthernet => 257,
-            HardwareType::Unknown(val) => val,
-        }
-    }
-}
-
-#[derive(Copy, Clone, Debug, PartialEq)]
-pub struct NetworkInterface {
-    pub interface_id: u32,
-    pub hardware_type: HardwareType,
-}
-
-#[derive(Copy, Clone, Debug, PartialEq)]
-pub enum StaticConfigType {
-    NoStaticConfig = 0x00,
-    StaticConfigIpv4 = 0x01,
-    StaticConfigIpv6 = 0x02,
-}
-
-impl TryFrom<u8> for StaticConfigType {
-    type Error = RdmError;
-
-    fn try_from(value: u8) -> Result<Self, Self::Error> {
-        match value {
-            0x00 => Ok(Self::NoStaticConfig),
-            0x01 => Ok(Self::StaticConfigIpv4),
-            0x02 => Ok(Self::StaticConfigIpv6),
-            value => Err(RdmError::InvalidStaticConfigType(value)),
-        }
-    }
-}
-
-#[derive(Copy, Clone, Debug, PartialEq)]
-pub enum BrokerState {
-    Disabled = 0x00,
-    Active = 0x01,
-    Standby = 0x02,
-}
-
-impl TryFrom<u8> for BrokerState {
-    type Error = RdmError;
-
-    fn try_from(value: u8) -> Result<Self, Self::Error> {
-        match value {
-            0x00 => Ok(Self::Disabled),
-            0x01 => Ok(Self::Active),
-            0x02 => Ok(Self::Standby),
-            value => Err(RdmError::InvalidBrokerState(value)),
-        }
-    }
-}
-
-#[derive(Copy, Clone, Debug, PartialEq)]
-pub enum DiscoveryState {
-    Incomplete,
-    Incremental,
-    Full,
-    NotActive,
-    ManufacturerSpecific(u8),
-}
-
-impl TryFrom<u8> for DiscoveryState {
-    type Error = RdmError;
-
-    fn try_from(value: u8) -> Result<Self, Self::Error> {
-        match value {
-            0x00 => Ok(Self::Incomplete),
-            0x01 => Ok(Self::Incremental),
-            0x02 => Ok(Self::Full),
-            0x04 => Ok(Self::NotActive),
-            value if (0x80..=0xff).contains(&value) => Ok(Self::ManufacturerSpecific(value)),
-            value => Err(RdmError::InvalidDiscoveryState(value)),
-        }
-    }
-}
-
-impl From<DiscoveryState> for u8 {
-    fn from(value: DiscoveryState) -> u8 {
-        match value {
-            DiscoveryState::Incomplete => 0x00,
-            DiscoveryState::Incremental => 0x01,
-            DiscoveryState::Full => 0x02,
-            DiscoveryState::NotActive => 0x04,
-            DiscoveryState::ManufacturerSpecific(value) => value,
-        }
-    }
-}
-
-#[derive(Copy, Clone, Debug, PartialEq)]
-pub enum DiscoveryCountStatus {
-    Incomplete,
-    Count(u16),
-    Unknown,
-}
-
-impl From<u16> for DiscoveryCountStatus {
-    fn from(value: u16) -> Self {
-        match value {
-            0x0000 => Self::Incomplete,
-            0xffff => Self::Unknown,
-            value => Self::Count(value),
-        }
-    }
-}
-
-impl From<DiscoveryCountStatus> for u16 {
-    fn from(value: DiscoveryCountStatus) -> u16 {
-        match value {
-            DiscoveryCountStatus::Incomplete => 0x0000,
-            DiscoveryCountStatus::Unknown => 0xffff,
-            DiscoveryCountStatus::Count(value) => value,
-        }
-    }
-}
-
-#[derive(Copy, Clone, Debug, PartialEq)]
-pub enum EndpointMode {
-    Disabled = 0x00, // Does not pass any DMX512-A/RDM traffic on a local RDM Command Port or DMX512-A Data Link
-    Input = 0x01, // Receives DMX512-A/RDM data on a local RDM Responder Port or DMX512-A Data Link
-    Output = 0x02, // Sends DMX512-A/RDM data on a local Command Port or DMX512-A Data Link
-}
-
-impl TryFrom<u8> for EndpointMode {
-    type Error = RdmError;
-
-    fn try_from(value: u8) -> Result<Self, Self::Error> {
-        match value {
-            0x00 => Ok(Self::Disabled),
-            0x01 => Ok(Self::Input),
-            0x02 => Ok(Self::Output),
-            value => Err(RdmError::InvalidEndpointMode(value)),
-        }
-    }
-}
-
-#[derive(Copy, Clone, Debug, PartialEq, Eq, Hash)]
-pub enum EndpointId {
-    Null,
-    Device(u16),
-    Reserved(u16),
-    Broadcast,
-}
-
-impl From<u16> for EndpointId {
-    fn from(value: u16) -> Self {
-        match value {
-            0 => EndpointId::Null,
-            0xffff => EndpointId::Broadcast,
-            value if (0xfa00..=0xfffe).contains(&value) => EndpointId::Reserved(value),
-            value => EndpointId::Device(value),
-        }
-    }
-}
-
-impl From<EndpointId> for u16 {
-    fn from(value: EndpointId) -> Self {
-        match value {
-            EndpointId::Broadcast => 0xffff,
-            EndpointId::Null => 0,
-            EndpointId::Device(value) => value,
-            EndpointId::Reserved(value) => value,
-        }
-    }
-}
-
-#[derive(Copy, Clone, Debug, PartialEq)]
-pub enum EndpointType {
-    Virtual = 0x00,
-    Physical = 0x01,
-}
-
-impl TryFrom<u8> for EndpointType {
-    type Error = RdmError;
-
-    fn try_from(value: u8) -> Result<Self, Self::Error> {
-        match value {
-            0x00 => Ok(Self::Virtual),
-            0x01 => Ok(Self::Physical),
-            value => Err(RdmError::InvalidEndpointType(value)),
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    #[cfg(feature = "alloc")]
-    fn should_decode_string_bytes() {
-        assert_eq!(
-            decode_string_bytes(&b"null terminated string\0"[..]).unwrap(),
-            "null terminated string".to_string()
-        );
-        assert_eq!(
-            decode_string_bytes(&b"not null terminated string"[..]).unwrap(),
-            "not null terminated string".to_string()
-        );
-        assert_eq!(
-            decode_string_bytes(&b"early terminated\0string"[..]).unwrap(),
-            "early terminated".to_string()
-        );
-    }
-
-    #[test]
-    #[cfg(not(feature = "alloc"))]
-    fn should_decode_string_bytes() {
-        assert_eq!(
-            decode_string_bytes::<32>(&b"null terminated string\0"[..]).unwrap(),
-            String::from_utf8(Vec::<u8, 32>::from_slice(b"null terminated string").unwrap())
-                .unwrap()
-        );
-        assert_eq!(
-            decode_string_bytes::<32>(&b"not null terminated string"[..]).unwrap(),
-            String::from_utf8(Vec::<u8, 32>::from_slice(b"not null terminated string").unwrap())
-                .unwrap()
-        );
-        assert_eq!(
-            decode_string_bytes::<32>(&b"early terminated\0string"[..]).unwrap(),
-            String::from_utf8(Vec::<u8, 32>::from_slice(b"early terminated").unwrap()).unwrap()
-        );
     }
 }
