+++ src/rdm/parameter/e133.rs
@@ -1,0 +1,41 @@
+use super::RdmError;
+
+#[derive(Copy, Clone, Debug, PartialEq)]
+pub enum StaticConfigType {
+    NoStaticConfig = 0x00,
+    StaticConfigIpv4 = 0x01,
+    StaticConfigIpv6 = 0x02,
+}
+
+impl TryFrom<u8> for StaticConfigType {
+    type Error = RdmError;
+
+    fn try_from(value: u8) -> Result<Self, Self::Error> {
+        match value {
+            0x00 => Ok(Self::NoStaticConfig),
+            0x01 => Ok(Self::StaticConfigIpv4),
+            0x02 => Ok(Self::StaticConfigIpv6),
+            value => Err(RdmError::InvalidStaticConfigType(value)),
+        }
+    }
+}
+
+#[derive(Copy, Clone, Debug, PartialEq)]
+pub enum BrokerState {
+    Disabled = 0x00,
+    Active = 0x01,
+    Standby = 0x02,
+}
+
+impl TryFrom<u8> for BrokerState {
+    type Error = RdmError;
+
+    fn try_from(value: u8) -> Result<Self, Self::Error> {
+        match value {
+            0x00 => Ok(Self::Disabled),
+            0x01 => Ok(Self::Active),
+            0x02 => Ok(Self::Standby),
+            value => Err(RdmError::InvalidBrokerState(value)),
+        }
+    }
+}
