+++ src/rdm/response.rs
@@ -48,13 +48,17 @@
 use super::{
     bsd_16_crc,
     parameter::{
-        decode_string_bytes, BrokerState, DefaultSlotValue, DhcpMode, DiscoveryCountStatus,
-        DiscoveryState, DisplayInvertMode, EndpointId, EndpointMode, EndpointType, Ipv4Address,
-        Ipv4Route, Ipv6Address, LampOnMode, LampState, MergeMode, NetworkInterface,
-        ParameterDescription, ParameterId, PinCode, PowerState, PresetPlaybackMode,
-        PresetProgrammed, ProductCategory, ProductDetail, ProtocolVersion, SelfTest,
-        SensorDefinition, SensorValue, SlotInfo, StaticConfigType, StatusMessage, StatusType,
-        SupportedTimes, TimeMode,
+        decode_string_bytes,
+        e120::{
+            DefaultSlotValue, DisplayInvertMode, LampOnMode, LampState, ParameterDescription,
+            PowerState, PresetPlaybackMode, ProductCategory, ProductDetail, ProtocolVersion,
+            SelfTest, SensorDefinition, SensorValue, SlotInfo, StatusMessage, StatusType,
+        },
+        e133::{BrokerState, StaticConfigType},
+        e137_1::{MergeMode, PinCode, PresetProgrammed, SupportedTimes, TimeMode},
+        e137_2::{DhcpMode, Ipv4Address, Ipv4Route, Ipv6Address, NetworkInterface},
+        e137_7::{DiscoveryCountStatus, DiscoveryState, EndpointId, EndpointMode, EndpointType},
+        ParameterId,
     },
     CommandClass, DeviceUID, RdmError, SubDeviceId, DISCOVERY_UNIQUE_BRANCH_PREAMBLE_BYTE,
     DISCOVERY_UNIQUE_BRANCH_PREAMBLE_SEPARATOR_BYTE, RDM_START_CODE_BYTE, RDM_SUB_START_CODE_BYTE,
